test(e2e): strengthen E2E assertions and remove legacy OpenCode test (#186)
